test(auth): Phase 0 characterization suite before broker unification
Pin current behavior of both auth implementations ahead of DRYing up the
standalone freeq-auth-broker and the embedded auth in freeq-server.

- Split freeq-auth-broker main.rs into lib.rs (pub router + helpers) and a
  thin binary so the HTTP surface is reachable from tests. Move-only; the
  3 existing unit tests pass unchanged.
- Add freeq-auth-broker/tests/characterization.rs (23 tests): /auth/callback
  fragment payload, known-nonce-first, nonce retry, mobile freeq:// 302,
  encrypted-at-rest storage, identity-only degrade; /session 401-vs-502
  classification, single-use refresh rotation, concurrent-refresh lock
  serialization, CSRF origin, scope defaulting; graph follow/unfollow
  authorization; HMAC push wire-format checked against the server verifier.
- Add freeq-server/tests/embedded_auth_callback.rs (4 tests) for the
  embedded-only paths: mobile redirect, irc_state /login completion, web
  session + one-time token minting.
- docs/AUTH-BROKER-UNIFICATION.md: survey, design, phasing, and findings.

Findings recorded, not fixed here: is_valid_return_to prefix-matches and
accepts irc.freeq.at.evil.example (residual C-6), pinned as a documented-bug
test; tests/evidence_verify.rs references a nonexistent freeq-verify bin
(pre-existing, unrelated).
